+++ docs/introduction/version_history.rst
@@ -12,6 +12,7 @@ Version 0.4.0 Beta (Not yet released)
 * Documentation fixes
 * Fixed compiling in RedHat/CentOS 6.x
 * Fixed parallel building
+* Fixed SSL connect with no server support (`Issue #68 <https://github.com/libattachsql/libattachsql/issues/68>`_)
 
 Version 0.3
 -----------